refactor: mistty--process-terminal-seq calls mistty--needs-refresh.
Before this change, many calls to (mistty--refresh)
had (mistty--needs-refresh) called just before to do a forced refresh,
to remain backward compatible.

With this change, the calls to (mistty--needs-refresh) are moved to
mistty--process-terminal-seq, since this is the reason why refresh is
needed. Additionally, refresh is skipped if the terminal seq is empty.