feat: Reduce blinking.
Before this change, the work buffer was refreshed after every single
string sent by the process attached to the terminal.

Depending on how the process buffers its output, this behaviour would
mean displaying confusing intermediate output. This typically happens
when refreshing: the process sends a string to delete the screen in one
call to write(), then redraws the screen in another call to write() made
right away. Emacs would then first clear the buffer then redraw it, and
that was visible.

This change attempts to avoid that by telling Emacs to read everything
it can from the process without waiting. The changes are buffered in the
term buffer then only copied into the work buffer when the process stops
sending data, so we don't see such intermediate state as the one
described below. We will still see intermediate state in the case where
there is a small pause between the two write(), even such as one caused
by a context switch.

Additionally, there is a very strict time limit in how much time Emacs
will spend accepting that process output, to avoid Emacs freezing when a
process continuously sends data to the terminal.