fix: Take point position from cursor when after inserted content.
In the common case where the point is after some inserted string, just
trust that the cursor will be at that position at the end of the replay
instead of relying on replace-buffer-contents to do its magic.

This is more reliable, as replace-buffer-contents magic fails sometimes,
especially when dealing with spaces, newlines or right prompts.

This change lacks a test, as I haven't been able to find a way to
trigger replace-buffer-contents magic to fail reliably. At least,
existing tests make sure that the change doesn't break existing
functionality.

Related to issue #39