fix: Use with-demoted-errors in pre and post command hook.
A function in pre or post command hook failing will result in that
function being removed from the hook. This would make a mistty buffer
completely nonfunctional.

Before this change, when there was an error in the mistty pre or post
command hook that function would be removed from the hook and the mistty
buffer would become unusable.

With this change, an error in the missty pre or post command hook
results in a message being printed, so at least the function stays.