fix: Turn off jit-lock-mode in the term buffer.
Before this change, it was possible for jit-lock-mode to apply
fontification either in the term or in the work buffer. This was ok as
long as fontification only modified text properties, as they're copied,
but goto-address-mode creates overlays.

This change explicitly turns off jit-lock-mode in term mode outside of
fullscreen mode, so fontification only ever happen in one buffer.

In practice, this resulted in unpredictable behavior with
goto-address-mode when enabled globally.

Issue #41