fix: Send BS instead of DEL to delete chars, because of Fish 4.0.0.
Before this change, MisTTY would send ^? a.k.a. \d a.k.a. \x7f a.k.a.
DEL to tell applications to execute backward-delete-char. This is one of
the two widely-used historical codes and seemed to work in the most
common cases.

Unfortunately, Fish 4.0.0 came out which by default handles ^? as delete
and deletes the char after point. This questionable move makes it so BS
wins as the most reasonable code to send.

This change can be undone if necessary by setting mistty-del to "\d" and
binding <backspace> to "\d" as well in mistty-term-map.