refactor: Simplify mistty-on-prompt.
Before this change, mistty-on-prompt tried to limit prompts detected
without relying on bracketed-paste to a single line, using
mistty-bracketed-paste.

This made the value of mistty-bracketed-paste overly important and
caused race conditions at the boundary, when mistty-bracketed-paste was
turned off in the term buffer but not yet in the work buffer.

This change removes this problematic and likely outdated piece of logic.
Suspiciously, this didn't require updating any tests.