feat: Use OSC 133 B to mark prompt fields.
With this change, when the shell sends out OSC 133 B to mark the end of
a prompt, MisTTY uses that information to mark the prompt like
shell-mode does. This way, functions like (line-beginning-position)
put the point at beginning of user input.

mistty--command-for-output uses the same mark to remove the prompt,
instead of relying on a regexp.

This makes it possible to use nonstandard prompts.