fix: Don't close prompt if end position is not accessible.
Before this change, MisTTY might try to close a prompt then get a nil
from mistty--scrollrow-pos, which would cause an error. This could happen
if the position where the prompt was closed was later on deleted.

This change checks the return value of mistty--scrollrow-pos before deciding
to close a prompt.