feat: Mark right-prompt as a field.
With this change, MisTTY lets Emacs know that right-prompts aren't part
of the command by adding the text property 'field 'right-prompt. This
way, end-of-line doesn't go past the right prompt.

This means that when OSC 113 B code is used, beginning-of-line and
end-of-line stay on the user input, as they should. Unfortunately,
without OSC 113 B code, MisTTY isn't confident enought about where
the (left) prompt starts to mark it the same way.