refactor: Remove mistty--set-sync-mark-from-end.
Before this change, updating the sync positions of the term and work
buffers was usually done counting chars from end, assuming that the two
buffer content were consistent.

This change switches to using scrollrows to sync the buffers. A specific
position is set on the work buffer and set to sync with the term buffer
starting at a given scrollrow. This should be safer, as it doesn't
require the two buffers to be consistent.