refactor: Reset and align based on scrollrow.
With this change, RESET works based on scrollrows instead of raw
positions. It tweaks the starting scrollrow of the terminal after the
reset so the scrollrows continue to increase and won't cause any
strange position problems.

Aligning the work and term buffer after a change on the terminal sync
mark is now done based on scrollrows as well: the sync mark of the work
buffer goes back to the scrollrow that correspond to the top of the
terminal. This is done by storing scrollrow in text properties when
moving the sync marker.