feat: Skip identical rows after realignment
Before this change, realignment would always move the sync marker back
to the beginning of the terminal screen. This changes puts the sync
marker back just before the last modified row.

This change also tests recovery with relevant sections of the work
buffer modified and tries to deal with that reasonably.