fix: Ignore fake newlines for scrollrow counts.
Before this change, terminal rows were sometimes used to compute
scrollrows. This wasn't correct, as two rows can become one if the
terminal is resized and a line was wrapped.

This change introduces primitives for working with scrollrows instead of
lines/rows:
 - go-beginning-of-scrollrow
 - beginning-of-scrollrow-pos (~ pos-bol)
 - go-end-of-scrollrow
 - end-of-scrollrow-pos (~pos-eol)
 - go-down-scrollrow (~ forward-line)
 - go-up-scrollrow

and rewrites code that work with scrollrows to use these functions.

This redefines a scrollrow as a terminal line (even if it is wrapped),
which should be stable even when the terminal is resized.