fix: Fix term-buffer-vertical-motion to allow arbitrary term size.
Before this change, term-buffer-vertical-motion was either:

 A. term.el implementation which tries to emulate vertical-motion but
 produces strange results.

 B. vertical-motion tied to mistty's window, or, failing that, the
 selected window, whose dimensions might or might not match term-width
 and term-height.

 C. vertical-motion tied to window showing the term-mode buffer
 in fullscreen mode

This change replaces A and B by forward-line, but leaves C in place.

Using forward-line is justified since term.el adds newlines where
wrapping should occur and doesn't rely on vertical-motion to do that.
This requires term-suppress-hard-newline to be nil, which this change
enforces.