test: Use a fixed terminal size by default.
Before this change, tests used a terminal size corresponding to the
selected window size.

This change switches most tests to a terminal size of 80x24, so it's
always the same. Tests that need a different terminal size or that need
the terminal size to track window site can specify it using the
:term-size key argument to mistty-with-test-buffer.