refactor: Let interacts communicate with the process
This change moves away from interacts returning a string to send to the
process. It instead lets interacts communicate with the process and
return when they want to wait for changes to be registered in the
process output.

This simplifies the queue, removing things like 'fire-and-forget, which
becomes encoded on the interact side by (mistty--interact-send) followed
by (mistty--interact-done) instead of (mistty--interact-return-then).