refactor: Add a (useless) accumulator in front of process filter.
This change introduces an accumulator type and puts it in front of the
process filter of any process created by mistty-exec. This accumulator
stays as process filter for the lifetime of the process, its output
redirected as needed to the relevant function.

The accumulator is meant to replace all of mistty custom process
filters, leaving only term's.