feat: Hold back chunk containing incomplete escape sequences.
With this change, incomplete escape sequences are detected by the
accumulator and held back until they're complete.

This change is based on the idea of generating regexp to detect
incomplete regexps automatically at compile-time when the processor is
created. This works based on the assumption that we can live with a
restricted subset of regular expressions for detecting terminal
sequence.

To make processing easier, regular expressions must now be written using
a limited subset of RX notation. This notation is expanded with
convenient shortcuts that map the notation used in the xterm control
sequence doc, such as ESC, CSI, OSC, Ps, Pn, Pt, ST.