fix: Make prompt-sp detection independent of chunking.
Before this change, prompt-sp detection relied on detecting the string
"     \r" in the current data, which might not be true when data is
chunked.

The normal hold-off logic of the accumulator was disabled for that
pattern because it wouldn't make sense to wait after every space for
maybe this pattern to appear.

This change switches to detecting just the \r and then looking back at
the data that was sent to the process filter to decide whether it's
worth it to flush and look for prompt-sp on the buffer.