fix: Copy rx-regexp in macros instead of modifying.
Before this change, mistty--accum-add-processor-lambda use c-copy to
allow in-place modifications of rx-regexp, to remove the let from one
and leave them for the other. This only worked if the let was at
toplevel, since cl-copy only does a shallow copy.

This change switches to copying the rx-notation tree instead of doing
in-place modifications. This avoids many potential issues.