refactor: Keep only functions in the oclosure, not data.
Keeping functions allows keeping data together with the rest of the
state of the accumulator. This simplifies the code without having to
move all the local functions definitions inside the oclosure.

This makes it possible for the list of processors to change after every
call to a processor. This function adapts the code for this.

Also fix typo: around, not arround.