fix: Limit unprocessed data that can be accumulated in bytes.
Before this change a time limit was set when accumulating immediately
pending data, 0.1s. While this isn't a long time to wait, this is enough
time to accumulate so much data that term.el takes several seconds to
process it.

This change introduces a size limit in addition to a time limit, so no
matter how fast data comes in, the accumulator won't accumulate so much
unprocessed data that it takes forever to process.