refactor: Copy mistty-bracketed-paste from term to work buffer.
Before this change, mistty-bracketed-paste was set on the work buffer
whenever it changed in the term buffer.

With this change, the variable is copied after a refresh, so the value
of the variable in the work buffer stays consistent with what's
displayed on the work buffer. This change also removes some unnecessary
special-casing of mistty-bracketed-paste.