fix: Only set prompt and right prompt field on the mistty buffer.
Before this change fields were set in the term-mode buffer, which
confused term.el, as it uses line-beginning-position and
line-end-position instead of pos-bol and pos-eol.

In practice, this resulted in any color in user input being erased.

With this change, MisTTY remembers the position of the start of user
input in prompt in the term mode buffer, then applies the field text
properties in the mistty mode buffer, on the line containing the start
of user input. This also avoids marking wrongly-detected right prompts,
not on the 1st line of user input.