fix: Avoid detecting right prompt containing or just after cursor.
Before this change, it was possible to get into a situation where the
cursor was inside the right prompt region or just at the border of it,
which caused strange behavior due to the right prompt field. This is
because MisTTY can't always know whether a whitespace after a command
is real (editable) or whether it was just added to move to the right
prompt.