feat: Store terminal data as escaped string in mistty-log.
Before this change, when data that's sent or received contained special
characters, they appeared in mistty-log as ^-notation, so for example ^[
for ESC. That was hard to read and inconsistent, so made debugging
harder.

With this change, data that's sent or received is formated as a string
using prin1, through (format "%S"), configured to escape special
characters, so now ESC appears as "\33".