refactor: interaction cb is only called once, returns the next cb.
Before this change, interactions were supposed to have a fixed callback
that was called over and over again, but:
 - this was normally not the case; each successful interaction would
   install a new callback
 - there was potential for loops, in issue #57

Wit this change, a callback is taken out of the interaction before
executing and another one is set afterwards. This way, we can't enter
into a loop if mistty--dequeue is called recursively.