fix: Call process-string at the end of mistty--dequeue only.
As shown in issue #57, calling process-string can trigger a call to the
process filter, if there's already output to process. Since interactions
send string and they are normally called from the process filter,
there's a loop, so mistty--decode could call itself recursively and see
some intermediate results.

To avoid that, mistty--interact-send now accumulate output into a list
that is then sent out by mistty--dequeue after it's done everything it
needed to do, in a state where recursive calls are acceptable.