fix: Avoid prompt_sp confusing prompt detection changing shells.
Before this change, it would be possible, when going from a shell with
prompt_sp enabled to a zsh with prompt_sp disabled, to end up in a
situation where the first prompt is seen as starting too early and be
discarded.

This change avoids the situation by systematically clearing the
prompt_sp mark of any detected bracketed paste or osc prompt.