fix: Keep process mark and point in sync in term-mode.
The term buffer normally keeps point at process mark and MisTTY relies
on that. The rule is that if the point is at process mark at the
beginning of term-emulate-terminal, it'll be put at process mark again
at the end.

However, when windows with long lines are resized, the point and process
mark end up being de-synced, and because of the rule above, they're
never re-synced. This surfaces in MisTTY as terminal zone not updating
anymore.

This change avoids the risk of the point and process mark de-syncing by
forcibly moving the point to the process mark after calling
term-emulate-terminal, which is always the right thing to do for MisTTY.

issue #63